Move calibration_dataset.py to backend/nxp/tests_models and install CifarNet example
